+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/GoogleIamDAO.scala
@@ -1,5 +1,7 @@
 package org.broadinstitute.dsde.workbench.google
 
+import ca.mrvisser.sealerate
+import org.broadinstitute.dsde.workbench.google.GoogleIamDAO.MemberType
 import org.broadinstitute.dsde.workbench.model._
 import org.broadinstitute.dsde.workbench.model.google._
 
@@ -80,7 +82,10 @@ trait GoogleIamDAO {
     * @param rolesToAdd Set of roles to add (example: roles/storage.admin)
     * @return true if the policy was updated; false otherwise.
     */
-  def addIamRolesForUser(iamProject: GoogleProject, email: WorkbenchEmail, rolesToAdd: Set[String]): Future[Boolean]
+  @deprecated(message = "Please use the generic method with {{{ memberType = MemberType.User }}}.", since = "0.21")
+  def addIamRolesForUser(iamProject: GoogleProject, email: WorkbenchEmail, rolesToAdd: Set[String]): Future[Boolean] = {
+    addIamRoles(iamProject: GoogleProject, email: WorkbenchEmail, MemberType.User, rolesToAdd: Set[String])
+  }
 
   /**
     * Removes project-level IAM roles for the given user.
@@ -91,7 +96,34 @@ trait GoogleIamDAO {
     * @param rolesToRemove Set of roles to remove (example: roles/dataproc.worker)
     * @return true if the policy was updated; false otherwise.
     */
-  def removeIamRolesForUser(iamProject: GoogleProject, email: WorkbenchEmail, rolesToRemove: Set[String]): Future[Boolean]
+  @deprecated(message = "Please use the generic method with {{{ memberType = MemberType.User }}}.", since = "0.21")
+  def removeIamRolesForUser(iamProject: GoogleProject, email: WorkbenchEmail, rolesToRemove: Set[String]): Future[Boolean] = {
+    removeIamRoles(iamProject: GoogleProject, email: WorkbenchEmail, MemberType.User, rolesToRemove: Set[String])
+  }
+
+  /**
+    * Adds project-level IAM roles for the given member type.
+    * This method will perform a read-modify-write of the project's IAM policy, and return a Boolean
+    * indicating whether a change was actually made.
+    * @param iamProject the project in which to add the roles
+    * @param email the email address
+    * @param memberType the type of member (e.g. 'user', 'group', 'service account')
+    * @param rolesToAdd Set of roles to add (example: roles/storage.admin)
+    * @return true if the policy was updated; false otherwise.
+    */
+  def addIamRoles(iamProject: GoogleProject, email: WorkbenchEmail, memberType: MemberType, rolesToAdd: Set[String]): Future[Boolean]
+
+  /**
+    * Removes project-level IAM roles for the given member type.
+    * This method will perform a read-modify-write of the project's IAM policy, and return a Boolean
+    * indicating whether a change was actually made.
+    * @param iamProject the google project in which to remove the roles
+    * @param email the email address
+    * @param memberType the type of member (e.g. 'user', 'group', 'service account')
+    * @param rolesToRemove Set of roles to remove (example: roles/dataproc.worker)
+    * @return true if the policy was updated; false otherwise.
+    */
+  def removeIamRoles(iamProject: GoogleProject, email: WorkbenchEmail, memberType: MemberType, rolesToRemove: Set[String]): Future[Boolean]
 
   /**
     * Adds the Service Account User role for the given users on the given service account.
@@ -134,4 +166,28 @@ trait GoogleIamDAO {
     * @return list of service account keys
     */
   def listUserManagedServiceAccountKeys(serviceAccountProject: GoogleProject, serviceAccountEmail: WorkbenchEmail): Future[Seq[ServiceAccountKey]]
+}
+
+object GoogleIamDAO {
+
+  /**
+    * Typing for the Google IAM member types as described at https://cloud.google.com/iam/docs/overview
+    */
+  sealed trait MemberType extends Serializable with Product
+  object MemberType {
+    final case object User extends MemberType {
+      override def toString = "user"
+    }
+    final case object Group extends MemberType {
+      override def toString = "group"
+    }
+    final case object ServiceAccount extends MemberType {
+      override def toString = "serviceAccount"
+    }
+    final case object Domain extends MemberType {
+      override def toString = "domain"
+    }
+
+    val stringToMemberType: Map[String, MemberType] = sealerate.collect[MemberType].map(p => (p.toString, p)).toMap
+  }
 }